Move JSON printer into TypedPDA class
